Split Image stories (#3387)
