+++ src/js/components/Image/stories/Fallback.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import { Grommet, Image } from 'grommet';
+import { grommet } from 'grommet/themes';
+
+const Fallback = () => (
+  <Grommet theme={grommet}>
+    <Image
+      fallback="//v2.grommet.io/assets/IMG_4245.jpg"
+      src="//v2.grommet.io/assets/IMG_4245_not_exists.jpg"
+    />
+  </Grommet>
+);
+
+storiesOf('Image', module).add('Fallback', () => <Fallback />);
